refactor: Move rec outside rmi module
Realm Execution Context (REC) has nothing to do with Realm Management
Interface (RMI). Thus, it would be better to locate Rec definition
outside rmi module. Note that mpidr, params, and run have not been
moved, as they represent `RmiRecMpidr`, `RmiRecParams`, `RmiRecRun`
respectively.

[before]
`struct Rec` defined in rmm/src/rmi/rec/mod.rs

[after]
`struct Rec` defined rmm/src/rec.rs
