+++ rmm/src/event/rsihandle.rs
@@ -6,8 +6,8 @@ use crate::get_granule;
 use crate::get_granule_if;
 use crate::granule::GranuleState;
 use crate::realm::rd::Rd;
+use crate::rec::Rec;
 use crate::rmi::rec::run::Run;
-use crate::rmi::rec::Rec;
 use crate::rsi;
 use crate::rsi::psci;
 use crate::Monitor;
