+++ rmm/src/rmi/rec/exit.rs
@@ -7,9 +7,9 @@ use crate::granule::GRANULE_MASK;
 use crate::realm::context::get_reg;
 use crate::realm::mm::stage2_tte::S2TTE;
 use crate::realm::rd::Rd;
+use crate::rec::Rec;
 use crate::rmi::error::Error;
 use crate::rmi::rec::run::Run;
-use crate::rmi::rec::Rec;
 use crate::rmi::rtt::is_protected_ipa;
 use crate::rmi::rtt::RTT_PAGE_LEVEL;
 use crate::Monitor;
